Fixing Perl Folding... Again
Case #1:

    sub test()
    {
        do_something(
            sub { },
        );
    }

Case #2:

    sub test()
    {
        do_something(
            sub {
            },
        );
    }

In both cases, the perlSubFold region pattern was permissive enough that
you would end up with 2 folds:

    1. From the start of test() to the terminating bracket
    2. From the 'sub {' line the terminating bracket of test()

This patch closes that opening. I'll grant that it only closes the issue
in these two specific cases, and may be too restrictive to be
comprehensive. On the other hand, these are common cases that I run into
in my employer's codebase. I imagine that they are probably the more
common cases elsewhere as well.